Add option to build shared lib and use fPIC
CMake is compiling each source file twice when building both static and shared lib. By adding an option we can build only either the shared or the static lib. A cleaner solution that allows building both at the same time is to use object libraries, but requires cmake 2.8.8 or later.