Seed using clocks instead of seconds
Using only time in seconds leads to the same seed for
all executions within the same second.

Clocks change faster so that this does not happen.
However, clock can give similar values for different the executions
of the program since it computes the time from the start
of the program. So we combine both time and clock.