Update main.cpp
For some reason, the output of the Eigen eigensolver does no longer order the eigenvalues in decreasing order (or it never did, and it is a problem since we replaced the ComplexEigenSolver by the regular EigenSolver). Added code to order the output such that the resulting translations make sense. This fixes both eigensolver and ge!